@@ -28,14 +28,14 @@ namespace Masuit.MyBlogs.Core.Controllers
     [ApiExplorerSettings(IgnoreApi = true)]
     public class UploadController : Controller
     {
-        public IHostingEnvironment HostingEnvironment { get; set; }
+        public IWebHostEnvironment HostEnvironment { get; set; }
 
         private readonly ImagebedClient _imagebedClient;
 
         /// <summary>
         /// 文件上传
         /// </summary>
-        /// <param name="hostingEnvironment"></param>
+        /// <param name="HostEnvironment"></param>
         public UploadController(ImagebedClient imagebedClient)
         {
             _imagebedClient = imagebedClient;
@@ -87,7 +87,7 @@ namespace Masuit.MyBlogs.Core.Controllers
                 return ResultData(null, false, "文件格式不支持，只能上传doc或者docx的文档!");
             }
 
-            string upload = HostingEnvironment.WebRootPath + "/upload";
+            string upload = HostEnvironment.WebRootPath + "/upload";
             if (!Directory.Exists(upload))
             {
                 Directory.CreateDirectory(upload);
@@ -139,7 +139,7 @@ namespace Masuit.MyBlogs.Core.Controllers
         public ActionResult Download(string path)
         {
             if (string.IsNullOrEmpty(path)) return Content("null");
-            var file = Path.Combine(HostingEnvironment.WebRootPath + "/upload", path.Trim('.', '/', '\\'));
+            var file = Path.Combine(HostEnvironment.WebRootPath + "/upload", path.Trim('.', '/', '\\'));
             if (System.IO.File.Exists(file))
             {
                 return this.ResumePhysicalFile(file, Path.GetFileName(file));
@@ -241,7 +241,7 @@ namespace Masuit.MyBlogs.Core.Controllers
                         return ResultData(url);
                     }
 
-                    path = Path.Combine(HostingEnvironment.WebRootPath, "upload", "images", filename);
+                    path = Path.Combine(HostEnvironment.WebRootPath, "upload", "images", filename);
                     var dir = Path.GetDirectoryName(path);
                     if (!Directory.Exists(dir))
                     {
@@ -255,13 +255,13 @@ namespace Masuit.MyBlogs.Core.Controllers
 
                     break;
                 case var _ when file.ContentType.StartsWith("audio") || file.ContentType.StartsWith("video"):
-                    path = Path.Combine(HostingEnvironment.WebRootPath, "upload", "media", filename);
+                    path = Path.Combine(HostEnvironment.WebRootPath, "upload", "media", filename);
                     break;
                 case var _ when file.ContentType.StartsWith("text") || (ContentType.Doc + "," + ContentType.Xls + "," + ContentType.Ppt + "," + ContentType.Pdf).Contains(file.ContentType):
-                    path = Path.Combine(HostingEnvironment.WebRootPath, "upload", "docs", filename);
+                    path = Path.Combine(HostEnvironment.WebRootPath, "upload", "docs", filename);
                     break;
                 default:
-                    path = Path.Combine(HostingEnvironment.WebRootPath, "upload", "files", filename);
+                    path = Path.Combine(HostEnvironment.WebRootPath, "upload", "files", filename);
                     break;
             }
             try
@@ -277,7 +277,7 @@ namespace Masuit.MyBlogs.Core.Controllers
                     file.CopyTo(fs);
                 }
 
-                return ResultData(path.Substring(HostingEnvironment.WebRootPath.Length).Replace("\\", "/"));
+                return ResultData(path.Substring(HostEnvironment.WebRootPath.Length).Replace("\\", "/"));
             }
             catch (Exception e)
             {
