@@ -15,7 +15,7 @@ namespace Masuit.MyBlogs.Core.Configs
         /// </summary>
         public static void Start()
         {
-            RecurringJob.AddOrUpdate(() => CheckLinks(), Cron.HourInterval(5)); //每5h检查友链
+            RecurringJob.AddOrUpdate(() => CheckLinks(), "0 */5 * * *"); //每5h检查友链
             RecurringJob.AddOrUpdate(() => EverydayJob(), Cron.Daily(5), TimeZoneInfo.Local); //每天的任务
             RecurringJob.AddOrUpdate(() => EveryweekJob(), Cron.Weekly(DayOfWeek.Monday, 5), TimeZoneInfo.Local); //每周的任务
             RecurringJob.AddOrUpdate(() => EveryHourJob(), Cron.Hourly); //每小时的任务
