@@ -3,7 +3,6 @@ using Masuit.MyBlogs.Core.Infrastructure.Services.Interface;
 using Masuit.MyBlogs.Core.Models.Enum;
 using Masuit.Tools.Logging;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Net.Http.Headers;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -32,7 +31,7 @@ namespace Masuit.MyBlogs.Core.Controllers
         /// 控制面板
         /// </summary>
         /// <returns></returns>
-        [Route("dashboard"), ResponseCache(Duration = 60, VaryByHeader = HeaderNames.Cookie)]
+        [Route("dashboard"), ResponseCache(Duration = 60, VaryByHeader = "Cookie")]
         public ActionResult Index()
         {
             return View();
@@ -119,7 +118,7 @@ namespace Masuit.MyBlogs.Core.Controllers
         /// 资源管理器
         /// </summary>
         /// <returns></returns>
-        [Route("filemanager"), ResponseCache(Duration = 60, VaryByHeader = HeaderNames.Cookie)]
+        [Route("filemanager"), ResponseCache(Duration = 60, VaryByHeader = "Cookie")]
         public ActionResult FileManager()
         {
             return View();
